TP2.2/Criação de componente Avatar e teste na página SingIn
