+++ src/components/avatar/index.jsx
@@ -1,0 +1,7 @@
+import { Avatar } from "@mui/material"
+
+const AvatarComponent = (props) => {
+    return <Avatar {...props} />;
+}
+
+export default AvatarComponent;
