+++ src/views/SignUp.jsx
